readme/TSP_example/requirements.txt
request == 2.22 . 0
json ==  2.0 . 9
itertools ==  8.0 . 2
typing ==  3.7 . 4.3